+++ src/Switchyard.Worker/IOrderingOutboxDispatchCycle.cs
@@ -1,0 +1,8 @@
+using Switchyard.Messaging;
+
+namespace Switchyard.Worker;
+
+public interface IOrderingOutboxDispatchCycle
+{
+    Task<OutboxDispatchResult> DispatchAsync(CancellationToken cancellationToken);
+}